projet_postgresql/serializers.py
# serializers.py

from rest_framework import serializers
from .models import Produit

class ProduitSerializer(serializers.ModelSerializer):
    class Meta:
        model = Produit
        fields = ['id', 'nom', 'description', 'prix', 'latitude', 'longitude', 'vector_search']
