requirements.txt
django==4.2.20
djangorestframework==3.15.2
psycopg2-binary==2.9.10
pgvector==0.3.6
python-decouple
